+++ backend/src/channel-invited/channel-invited.resolver.ts
@@ -1,4 +1,11 @@
-import { Args, Mutation, Resolver, Query, Subscription } from '@nestjs/graphql'
+import {
+  Args,
+  Mutation,
+  Resolver,
+  Query,
+  Subscription,
+  Context
+} from '@nestjs/graphql'
 import { ChannelInvited } from './entities/channel-invited.entity'
 import { ChannelInvitedService } from './channel-invited.service'
 import { CreateChannelInvitedInput } from './dto/create-channel-invited.input'
@@ -9,13 +16,20 @@ import {
   Unprotected
 } from '../auth/guards/authorization.guard'
 import { PubSub } from 'graphql-subscriptions'
+import { PrismaService } from 'src/prisma/prisma.service'
+import {
+  ForbiddenAccessChannelAdmin,
+  channelAdminGuard,
+  userContextGuard
+} from 'src/auth/guards/request.guards'
 
 @Resolver(() => ChannelInvited)
 @UseGuards(AuthorizationGuard)
 export class ChannelInvitedResolver {
   constructor(
     private readonly channelInvitedService: ChannelInvitedService,
-    private readonly pubSub: PubSub
+    private readonly pubSub: PubSub,
+    private prisma: PrismaService
   ) {}
 
   //**************************************************//
@@ -51,9 +65,32 @@ export class ChannelInvitedResolver {
   @Mutation(() => ChannelInvited)
   async createChannelInvited(
     @Args('data', { type: () => CreateChannelInvitedInput }, ValidationPipe)
-    data: CreateChannelInvitedInput
+    data: CreateChannelInvitedInput,
+    @Context() ctx: any
   ): Promise<ChannelInvited> {
-    return this.channelInvitedService.create(data)
+    if (
+      !(await channelAdminGuard(
+        ctx?.req?.user?.id,
+        data.channelId,
+        this.prisma
+      ))
+    )
+      throw new ForbiddenAccessChannelAdmin()
+    const res = await this.channelInvitedService.create(data)
+
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId: data.channelId }
+    })
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('invitationReceived-' + value.userId, { res })
+      })
+    )
+
+    await this.pubSub.publish('invitationReceived-' + data.userId, { res })
+
+    return res
   }
 
   @Mutation(() => ChannelInvited)
@@ -61,9 +98,31 @@ export class ChannelInvitedResolver {
     @Args('userId', { type: () => String }, NanoidValidationPipe)
     userId: string,
     @Args('channelId', { type: () => String }, NanoidValidationPipe)
-    channelId: string
+    channelId: string,
+    @Context() ctx: any
   ): Promise<ChannelInvited> {
-    return this.channelInvitedService.delete(userId, channelId)
+    console.log('userId:', userId)
+    console.log('userIdcontext:', ctx?.req?.user?.id)
+    if (
+      !(await channelAdminGuard(ctx?.req?.user?.id, channelId, this.prisma)) &&
+      !userContextGuard(ctx?.req?.user?.id, userId)
+    )
+      throw new ForbiddenAccessChannelAdmin()
+    const res = await this.channelInvitedService.delete(userId, channelId)
+
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId }
+    })
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('invitationDeleted-' + value.userId, { res })
+      })
+    )
+
+    await this.pubSub.publish('invitationDeleted-' + userId, { res })
+
+    return res
   }
 
   //**************************************************//
