+++ backend/src/channel-member/channel-member.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable } from '@nestjs/common'
 import { PrismaService } from '../prisma/prisma.service'
-import { ChannelMember, EChannelType, EMemberType } from '@prisma/client'
+import { ChannelMember, EMemberType } from '@prisma/client'
 import { CreateChannelMemberInput } from './dto/create-channel-member.input'
 import { UpdateChannelMemberInput } from './dto/update-channel-member.input'
 import { ChannelBlockedService } from '../channel-blocked/channel-blocked.service'
@@ -8,7 +8,6 @@ import { ChannelInvitedService } from '../channel-invited/channel-invited.servic
 import { ChannelService } from '../channel/channel.service'
 import { UserService } from '../user/user.service'
 import { ExceptionUserBlockedInChannel } from '../channel/exceptions/blocked.exception'
-import { ExceptionUserNotInvited } from '../channel/exceptions/invited.exception'
 import { ExceptionMaxUserReachedInChannel } from '../channel/exceptions/channel.exception'
 import {
   ExceptionTryingToMakeAdminAnAdmin,
@@ -53,28 +52,25 @@ export class ChannelMemberService {
     const numberMembers = (
       await this.prisma.channelMember.findMany({ where: { channelId } })
     ).length
+    const user = await this.userService.findOne(data.userId)
+
+    if (!user) throw new ExceptionUserNotFound()
 
     if (userBlocked) {
       throw new ExceptionUserBlockedInChannel()
     }
-    if (channel && channel.type === EChannelType.Protected) {
-      if (userInvited)
-        await this.channelInvitedService.delete(userId, channelId)
-      else throw new ExceptionUserNotInvited()
-    }
+
+    if (userInvited) await this.channelInvitedService.delete(userId, channelId)
+
     if (channel && numberMembers >= channel.maxUsers) {
       throw new ExceptionMaxUserReachedInChannel()
     }
 
-    let nickname = data.nickname
-    if (!nickname) {
-      const user = await this.userService.findOne(data.userId)
-      if (!user) throw new ExceptionUserNotFound()
-      nickname = user.username
-    }
+    const nickname = data.nickname || user.username
+    const avatarUrl = data.avatarUrl || user.avatarUrl
 
     return this.prisma.channelMember.create({
-      data: { nickname, ...data }
+      data: { nickname, avatarUrl, ...data }
     })
   }
