+++ backend/src/channel/channel.resolver.ts
@@ -1,4 +1,11 @@
-import { Resolver, Query, Mutation, Args, Subscription } from '@nestjs/graphql'
+import {
+  Resolver,
+  Query,
+  Mutation,
+  Args,
+  Subscription,
+  Context
+} from '@nestjs/graphql'
 import { ChannelService } from './channel.service'
 import { Channel } from './entities/channel.entity'
 import { CreateChannelInput } from './dto/create-channel.input'
@@ -16,13 +23,24 @@ import {
 import { NanoidsValidationPipe } from 'src/common/pipes/nanoids.pipe'
 import { UsernameValidationPipe } from 'src/common/pipes/username.pipe'
 import { PubSub } from 'graphql-subscriptions'
+import { PrismaService } from 'src/prisma/prisma.service'
+import {
+  ForbiddenAccessChannelAdmin,
+  ForbiddenAccessChannelOwner,
+  ForbiddenAccessData,
+  channelAdminGuard,
+  channelOwnerGuard,
+  userContextGuard
+} from 'src/auth/guards/request.guards'
+import * as bcrypt from 'bcrypt'
 
 @Resolver(() => Channel)
 @UseGuards(AuthorizationGuard)
 export class ChannelResolver {
   constructor(
     private readonly channelService: ChannelService,
-    private readonly pubSub: PubSub
+    private readonly pubSub: PubSub,
+    private prisma: PrismaService
   ) {}
 
   //**************************************************//
@@ -58,8 +76,12 @@ export class ChannelResolver {
   @Mutation(() => Channel)
   async createChannel(
     @Args('data', { type: () => CreateChannelInput }, ValidationPipe)
-    data: CreateChannelInput
+    data: CreateChannelInput,
+    @Context() ctx: any
   ): Promise<Channel> {
+    if (!userContextGuard(ctx?.req?.user?.id, data.ownerId))
+      throw new ForbiddenAccessData()
+    if (data?.password) data.password = bcrypt.hashSync(data.password, 10)
     return this.channelService.create(data)
   }
 
@@ -68,24 +90,73 @@ export class ChannelResolver {
     @Args('id', { type: () => String }, NanoidValidationPipe)
     id: string,
     @Args('data', { type: () => UpdateChannelInput }, ValidationPipe)
-    data: UpdateChannelInput
+    data: UpdateChannelInput,
+    @Context() ctx: any
   ): Promise<Channel> {
-    return this.channelService.update(id, data)
+    if (!(await channelAdminGuard(ctx?.req?.user?.id, id, this.prisma)))
+      throw new ForbiddenAccessChannelAdmin()
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId: id }
+    })
+
+    const res = await this.channelService.update(id, data)
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('channelEdited-' + value.userId, { res })
+      })
+    )
+
+    return res
   }
 
   @Mutation(() => Channel)
   async updateChannelOwner(
+    @Args('id', { type: () => String }, NanoidValidationPipe)
     id: string,
-    data: UpdateChannelOwnerIdInput
+    @Args('data', { type: () => UpdateChannelOwnerIdInput }, ValidationPipe)
+    data: UpdateChannelOwnerIdInput,
+    @Context() ctx: any
   ): Promise<Channel | null> {
-    return this.channelService.updateOwner(id, data)
+    if (!(await channelOwnerGuard(ctx?.req?.user?.id, id, this.prisma)))
+      throw new ForbiddenAccessChannelOwner()
+
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId: id }
+    })
+
+    const res = await this.channelService.updateOwner(id, data)
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('channelEdited-' + value.userId, { res })
+      })
+    )
+
+    return res
   }
 
   @Mutation(() => Channel)
   async deleteChannel(
-    @Args('id', { type: () => String }, NanoidValidationPipe) id: string
+    @Args('id', { type: () => String }, NanoidValidationPipe) id: string,
+    @Context() ctx: any
   ): Promise<Channel> {
-    return this.channelService.delete(id)
+    if (!(await channelOwnerGuard(ctx?.req?.user?.id, id, this.prisma)))
+      throw new ForbiddenAccessChannelOwner()
+
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId: id }
+    })
+
+    const res = await this.channelService.delete(id)
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('channelDeleted-' + value.userId, { res })
+      })
+    )
+
+    return res
   }
 
   //**************************************************//
